flag *.c extension on input files (but not *.bpf.c)Show description for f4b52e4
Users should be notified when they dont use the *.bpf.c standard. Originally, the solution was to modify
the library interfaces, and treat *.bpf.c violation as a "lint", conforming to the current project workflow.
However, this requires a function signature change since lint() only takes code and code does not contain filename
metadata. As such, it was decided a simple solution for now was to perform a check on the cli interface srcs and warn
whenever a non *.bpf.c file is detected.

importantly:

/cli/src/main.rs now contains a function to verify correct file usage, in addition to its unit tests.

/src/report.rs now contains a special check in report_terminal() for empty byte ranges, hiding line output on empty ranges. In addition, a unit test was added to test empty byte range functionality.