Add workflow for updating tree-sitter-0.21 branch
For <reasons> we need to have a version of the library that works with
tree-sitter version 0.21. To that end we have a branch, tree-sitter-0.21
that contains a patch set making this happen. However, rebasing said
patch set is tedious and, contrary to "regular" workflows, we should not
just rebase and throw away old commits but keep them around.
This change adds a workflow that automates this process: whenever a
change is pushed to main, it will merge these changes into the
tree-sitter-0.21 branch. As a result, we either get instant updates of
said branch, ready for use, or an immediate notification that the merge
did not happen successfully, indicating that the patch set needs to be
adjusted.