Add probe-read test for divergent signature
C doesn't support functions, but it's still conceivable (although highly
unlikely) that a function with the name bpf_probe_read is defined with a
different signature of the BPF provided one (while being careful to not
include the corresponding BPF header).
Add a test to make sure that we do not accidentally lint such
mismatching candidates.