Use workspace-wide definitions for various metadata
Cargo now allows us to define certain Cargo metadata in a workspace-wide
manner and then more easily reuse that. We intend to update things such
as the edition for all crates in the workspace. This makes those
properties (and others) perfect candidates for this infrastructure.
With this change we switch over to using workspace-wide properties for a
bunch of metadata.