Introduce Lint::message member
So far the lint message to be displayed to the user on match has been
provided as a property of the tree-sitter query itself. This is becoming
more and more of a burden. One of the minor issues is that there is
absolutely no support for splitting a line across multiple lines, so
longer messages and up being cumbersome to work with in the .scm file.
More severely, though, by having messages co-located with the query code
itself, it is virtually impossible to overwrite them. Lastly, because a
query does not have to contain such a message, we enabled additional
failure paths if such a message is not available.
To fix these problems, this change moves the message into a separate
property inside the Lint type. In terms of file system representation,
we simply co-locate a .txt file containing the message with the .scm
file containing the query code.