Make wasm-bindgen-cli-support build-dependency optional
Target specifications as we use them for wasm-bindgen-cli-support don't
really work for build script, because those are build for the host (as
opposed to the target) anyway. The only reason said dependency was
found, was because we have it specified as a regular dependency anyway.
Remove the target specific definition and then guard it by the 'deploy'
feature to truly make it optional.