Fix up 'cargo deploy' infrastructure
As it turns out, the 'cargo deploy' alias doesn't work properly when the
target/ directory is empty. The reason for that is that it doesn't have
the same semantics as the Makefile based logic, which always required a
regular 'cargo build' before doing a "deploy".
This change fixes up this problem, by adjusting the alias to execute
said build first. To make that happen, we generalize the 'serve' crate
into 'make' and add support for running arbitrary shell commands.