Use smaller logo in README
Our README is optimized for viewing on GitHub as its the most relevant,
but unfortunately the crates.io Markdown renderer produces a much more
ugly result, with the height styling effectively being ignored and the
logo displayed at full size.
To make that less of a thorn, add a smaller logo variant and use that.