Add support for highlighting in output
Add a new CLI option `--color` that enables 24bit color in the console.
This support is enabled via tree-sitter highlight queries.

The original queries are hydrated from the tree-sitter-c grammar with an
additional eBPF query for `SEC` for demonstrative purposes.

At the moment there is no support for theming and a single color scheme
is provided.