Rename COLOR_BLUE to COLOR_INDIGO
In the future we are likely going to expand our set of colors. At the
same time we want to stay close to the GitHub Sublime theme that we had
been using. One of the conflicts we anticipate is for the color blue,
for which we would also like to use the terminal based definition moving
forward.
To prevent any conflicts from showing up, rename the existing COLOR_BLUE
to COLOR_INDIGO.